refactor: Header & SideBar 코드 리뷰 반영
