+++ src/assets/svgs/header/BtnMenu.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgBtnMenu = (props: SVGProps<SVGSVGElement>) => (
   <svg
