+++ src/assets/svgs/header/CmpBtnBack.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgCmpBtnBack = (props: SVGProps<SVGSVGElement>) => (
   <svg
