+++ src/assets/svgs/header/ImgHeaderLogo.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgImgHeaderLogo = (props: SVGProps<SVGSVGElement>) => (
   <svg
