+++ src/assets/svgs/sidebar/BtnAccordion.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgBtnAccordion = (props: SVGProps<SVGSVGElement>) => (
   <svg
