+++ src/assets/svgs/sidebar/BtnX.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgBtnX = (props: SVGProps<SVGSVGElement>) => (
   <svg
