+++ src/assets/svgs/sidebar/IcSearch.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgIcSearch = (props: SVGProps<SVGSVGElement>) => (
   <svg
