+++ src/constants/sidebarList.ts
@@ -1,0 +1,30 @@
+export const SIDEBAR_LIST = [
+  {
+    title: 'Menu',
+    subItems: [
+      '버거',
+      '맥런치',
+      '맥모닝',
+      '해피 스낵',
+      '사이드&디저트',
+      '맥카페&음료',
+      '해피밀',
+    ],
+  },
+  {
+    title: 'Store',
+    subItems: ['매장찾기', '맥드라이브', '맥딜리버리', '임차문의'],
+  },
+  {
+    title: "What's New",
+    subItems: ['프로모션', '새로운 소식', '이달의 해피밀'],
+  },
+  {
+    title: 'Brand Story',
+    subItems: ['스토리', '브랜드소개', '맥도날드 사람들', '사회적 책임과 지원'],
+  },
+  {
+    title: 'Food Quality Story',
+    subItems: [],
+  },
+];
